server/pyproject.toml
[project]
name = "halloween-server"
version = "0.1.0"
description = "Halloween Haunted House MQTT Control Server"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=1.6.1",
    "matplotlib>=3.8.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
]
